feat(tabs): rename tabs.module.css => style.module.css
